Fix class file access problems during JUnit tests
During JUnit tests on - potentially fast - systems, a class file access
problem may occur with the compiled BASIC class file. This is fixed by
creating a temporary folder containg the compiled BASIC class file. The
temporary folder and the class file are deleted after each test. In
addition, File.delete() does not seem to delete a folder reliably, so
deleteOnExit() is applied on the temporary folder to ensure that it is
deleted when the JVM exits.